V2 Stream methods in AppFabric handler to forward to V3 AppFabricStreamHttpHandler
