+++ cdap-app-fabric/src/main/java/co/cask/cdap/gateway/handlers/AppFabricStreamHttpHandler.java
@@ -1,0 +1,134 @@
+/*
+ * Copyright © 2015 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package co.cask.cdap.gateway.handlers;
+
+import co.cask.cdap.app.services.Data;
+import co.cask.cdap.app.store.Store;
+import co.cask.cdap.app.store.StoreFactory;
+import co.cask.cdap.common.conf.CConfiguration;
+import co.cask.cdap.common.conf.Constants;
+import co.cask.cdap.data.Namespace;
+import co.cask.cdap.data2.datafabric.DefaultDatasetNamespace;
+import co.cask.cdap.data2.dataset2.DatasetFramework;
+import co.cask.cdap.data2.dataset2.NamespacedDatasetFramework;
+import co.cask.cdap.data2.transaction.stream.StreamAdmin;
+import co.cask.cdap.gateway.auth.Authenticator;
+import co.cask.cdap.gateway.handlers.util.AbstractAppFabricHttpHandler;
+import co.cask.cdap.proto.ProgramType;
+import co.cask.http.HttpResponder;
+import com.google.inject.Inject;
+import org.jboss.netty.handler.codec.http.HttpRequest;
+import org.jboss.netty.handler.codec.http.HttpResponseStatus;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.ws.rs.DELETE;
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+
+/**
+ *  HttpHandler class for app-fabric requests.
+ */
+@Path(Constants.Gateway.API_VERSION_3 + "/namespaces/{namespace-id}")
+public class AppFabricStreamHttpHandler extends AbstractAppFabricHttpHandler {
+  private static final Logger LOG = LoggerFactory.getLogger(AppFabricStreamHttpHandler.class);
+
+  /**
+   * Access Dataset Service
+   */
+  private final DatasetFramework dsFramework;
+
+  /**
+   * Store manages non-runtime lifecycle.
+   */
+  private final Store store;
+
+
+  private final StreamAdmin streamAdmin;
+
+
+  /**
+   * Constructs an new instance. Parameters are binded by Guice.
+   */
+  @Inject
+  public AppFabricStreamHttpHandler(Authenticator authenticator, CConfiguration configuration,
+                                    StoreFactory storeFactory, StreamAdmin streamAdmin, DatasetFramework dsFramework) {
+
+    super(authenticator);
+    this.streamAdmin = streamAdmin;
+    this.store = storeFactory.create();
+    this.dsFramework =
+      new NamespacedDatasetFramework(dsFramework, new DefaultDatasetNamespace(configuration, Namespace.USER));
+  }
+
+  @DELETE
+  @Path("/streams")
+  public void deleteStreams(HttpRequest request, HttpResponder responder,
+                            @PathParam("namespace-id") String namespaceId) {
+    try {
+      streamAdmin.dropAll();
+      responder.sendStatus(HttpResponseStatus.OK);
+    } catch (Exception e) {
+      LOG.error("Error while deleting streams", e);
+      responder.sendString(HttpResponseStatus.INTERNAL_SERVER_ERROR, e.getMessage());
+    }
+  }
+
+  /**
+   * Returns a list of streams associated with account.
+   */
+  @GET
+  @Path("/streams")
+  public void getStreams(HttpRequest request, HttpResponder responder,
+                         @PathParam("namespace-id") String namespaceId) {
+    dataList(request, responder, store, dsFramework, Data.STREAM, null, null);
+  }
+
+  /**
+   * Returns a stream associated with account.
+   */
+  @GET
+  @Path("/streams/{stream-id}")
+  public void getStreamSpecification(HttpRequest request, HttpResponder responder,
+                                     @PathParam("namespace-id") String namespaceId,
+                                     @PathParam("stream-id") final String streamId) {
+    dataList(request, responder, store, dsFramework, Data.STREAM, streamId, null);
+  }
+
+  /**
+   * Returns a list of streams associated with application.
+   */
+  @GET
+  @Path("/apps/{app-id}/streams")
+  public void getStreamsByApp(HttpRequest request, HttpResponder responder,
+                              @PathParam("namespace-id") String namespaceId,
+                              @PathParam("app-id") final String appId) {
+    dataList(request, responder, store, dsFramework, Data.STREAM, null, appId);
+  }
+
+  /**
+   * Returns all flows associated with a stream.
+   */
+  @GET
+  @Path("/streams/{stream-id}/flows")
+  public void getFlowsByStream(HttpRequest request, HttpResponder responder,
+                               @PathParam("namespace-id") String namespaceId,
+                               @PathParam("stream-id") final String streamId) {
+    programListByDataAccess(request, responder, store, dsFramework, ProgramType.FLOW, Data.STREAM, streamId);
+  }
+}
