+++ cdap-app-fabric/src/main/java/co/cask/cdap/gateway/handlers/util/AbstractAppFabricHttpHandler.java
@@ -17,30 +17,44 @@
 package co.cask.cdap.gateway.handlers.util;
 
 import co.cask.cdap.api.ProgramSpecification;
+import co.cask.cdap.api.data.stream.StreamSpecification;
+import co.cask.cdap.api.dataset.DatasetSpecification;
+import co.cask.cdap.api.flow.FlowSpecification;
+import co.cask.cdap.api.flow.FlowletConnection;
+import co.cask.cdap.api.flow.FlowletDefinition;
+import co.cask.cdap.api.mapreduce.MapReduceSpecification;
+import co.cask.cdap.api.procedure.ProcedureSpecification;
 import co.cask.cdap.app.ApplicationSpecification;
 import co.cask.cdap.app.runtime.ProgramRuntimeService;
+import co.cask.cdap.app.services.Data;
 import co.cask.cdap.app.store.Store;
 import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.http.RESTMigrationUtils;
+import co.cask.cdap.data2.dataset2.DatasetFramework;
 import co.cask.cdap.gateway.auth.Authenticator;
 import co.cask.cdap.gateway.handlers.AuthenticatedHttpHandler;
 import co.cask.cdap.internal.UserErrors;
 import co.cask.cdap.internal.UserMessages;
+import co.cask.cdap.proto.DatasetRecord;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.Instances;
 import co.cask.cdap.proto.ProgramRecord;
 import co.cask.cdap.proto.ProgramType;
+import co.cask.cdap.proto.StreamRecord;
 import co.cask.http.HttpResponder;
 import com.google.common.base.Charsets;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.Lists;
+import com.google.common.collect.Sets;
 import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
 import com.google.gson.reflect.TypeToken;
 import org.jboss.netty.buffer.ChannelBuffer;
 import org.jboss.netty.buffer.ChannelBufferInputStream;
+import org.jboss.netty.handler.codec.http.HttpHeaders;
 import org.jboss.netty.handler.codec.http.HttpRequest;
 import org.jboss.netty.handler.codec.http.HttpResponseStatus;
 import org.slf4j.Logger;
@@ -54,6 +68,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
+import java.util.Set;
 import javax.annotation.Nullable;
 
 /**
@@ -326,5 +341,251 @@ public abstract class AbstractAppFabricHttpHandler extends AuthenticatedHttpHand
     request.setUri(originalUri.replaceFirst(Constants.Gateway.API_VERSION_2, Constants.Gateway.API_VERSION_3 +
       "/namespaces/" + Constants.DEFAULT_NAMESPACE));
     return request;
+  }
+
+
+  protected final void dataList(HttpRequest request, HttpResponder responder, Store store, DatasetFramework dsFramework,
+                        Data type, String name, String appId) {
+    try {
+      if ((name != null && name.isEmpty()) || (appId != null && appId.isEmpty())) {
+        responder.sendString(HttpResponseStatus.BAD_REQUEST, "Empty name provided");
+        return;
+      }
+
+      String accountId = getAuthenticatedAccountId(request);
+      Id.Program program = Id.Program.from(accountId, appId == null ? "" : appId, "");
+      String json = name != null ? getDataEntity(store, dsFramework, program, type, name) :
+        appId != null ? listDataEntitiesByApp(store, dsFramework, program, type)
+          : listDataEntities(store, dsFramework, program, type);
+      if (json.isEmpty()) {
+        responder.sendStatus(HttpResponseStatus.NOT_FOUND);
+      } else {
+        responder.sendByteArray(HttpResponseStatus.OK, json.getBytes(Charsets.UTF_8),
+                                ImmutableMultimap.of(HttpHeaders.Names.CONTENT_TYPE, "application/json"));
+      }
+    } catch (SecurityException e) {
+      responder.sendStatus(HttpResponseStatus.UNAUTHORIZED);
+    } catch (Throwable e) {
+      LOG.error("Got exception : ", e);
+      responder.sendStatus(HttpResponseStatus.INTERNAL_SERVER_ERROR);
+    }
+  }
+
+  private String getDataEntity(Store store, DatasetFramework dsFramework,
+                               Id.Program programId, Data type, String name) {
+    Id.Namespace namespace = new Id.Namespace(programId.getNamespaceId());
+    if (type == Data.DATASET) {
+      DatasetSpecification dsSpec = getDatasetSpec(dsFramework, name);
+      String typeName = null;
+      if (dsSpec != null) {
+        typeName = dsSpec.getType();
+      }
+      return GSON.toJson(makeDataSetRecord(name, typeName));
+    } else if (type == Data.STREAM) {
+      StreamSpecification spec = store.getStream(namespace, name);
+      return spec == null ? "" : GSON.toJson(makeStreamRecord(spec.getName(), spec));
+    }
+    return "";
+  }
+
+  private String listDataEntities(Store store, DatasetFramework dsFramework,
+                                  Id.Program programId, Data type) throws Exception {
+    if (type == Data.DATASET) {
+      Collection<DatasetSpecification> instances = dsFramework.getInstances();
+      List<DatasetRecord> result = Lists.newArrayListWithExpectedSize(instances.size());
+      for (DatasetSpecification instance : instances) {
+        result.add(makeDataSetRecord(instance.getName(), instance.getType()));
+      }
+      return GSON.toJson(result);
+    } else if (type == Data.STREAM) {
+      Collection<StreamSpecification> specs = store.getAllStreams(new Id.Namespace(programId.getNamespaceId()));
+      List<StreamRecord> result = Lists.newArrayListWithExpectedSize(specs.size());
+      for (StreamSpecification spec : specs) {
+        result.add(makeStreamRecord(spec.getName(), null));
+      }
+      return GSON.toJson(result);
+    }
+    return "";
+
+  }
+
+  private String listDataEntitiesByApp(Store store, DatasetFramework dsFramework,
+                                       Id.Program programId, Data type) throws Exception {
+    Id.Namespace namespace = new Id.Namespace(programId.getNamespaceId());
+    ApplicationSpecification appSpec = store.getApplication(new Id.Application(
+      namespace, programId.getApplicationId()));
+    if (type == Data.DATASET) {
+      Set<String> dataSetsUsed = dataSetsUsedBy(appSpec);
+      List<DatasetRecord> result = Lists.newArrayListWithExpectedSize(dataSetsUsed.size());
+      for (String dsName : dataSetsUsed) {
+        String typeName = null;
+        DatasetSpecification dsSpec = getDatasetSpec(dsFramework, dsName);
+        if (dsSpec != null) {
+          typeName = dsSpec.getType();
+        }
+        result.add(makeDataSetRecord(dsName, typeName));
+      }
+      return GSON.toJson(result);
+    }
+    if (type == Data.STREAM) {
+      Set<String> streamsUsed = streamsUsedBy(appSpec);
+      List<StreamRecord> result = Lists.newArrayListWithExpectedSize(streamsUsed.size());
+      for (String streamName : streamsUsed) {
+        result.add(makeStreamRecord(streamName, null));
+      }
+      return GSON.toJson(result);
+    }
+    return "";
+  }
+
+  @Nullable
+  private DatasetSpecification getDatasetSpec(DatasetFramework dsFramework, String dsName) {
+    try {
+      return dsFramework.getDatasetSpec(dsName);
+    } catch (Exception e) {
+      LOG.warn("Couldn't get spec for dataset: " + dsName);
+      return null;
+    }
+  }
+
+  private Set<String> dataSetsUsedBy(FlowSpecification flowSpec) {
+    Set<String> result = Sets.newHashSet();
+    for (FlowletDefinition flowlet : flowSpec.getFlowlets().values()) {
+      result.addAll(flowlet.getDatasets());
+    }
+    return result;
+  }
+
+  private Set<String> dataSetsUsedBy(ApplicationSpecification appSpec) {
+    Set<String> result = Sets.newHashSet();
+    for (FlowSpecification flowSpec : appSpec.getFlows().values()) {
+      result.addAll(dataSetsUsedBy(flowSpec));
+    }
+    for (ProcedureSpecification procSpec : appSpec.getProcedures().values()) {
+      result.addAll(procSpec.getDataSets());
+    }
+    for (MapReduceSpecification mrSpec : appSpec.getMapReduce().values()) {
+      result.addAll(mrSpec.getDataSets());
+    }
+    return result;
+  }
+
+  private Set<String> streamsUsedBy(FlowSpecification flowSpec) {
+    Set<String> result = Sets.newHashSet();
+    for (FlowletConnection con : flowSpec.getConnections()) {
+      if (FlowletConnection.Type.STREAM == con.getSourceType()) {
+        result.add(con.getSourceName());
+      }
+    }
+    return result;
+  }
+
+  private Set<String> streamsUsedBy(ApplicationSpecification appSpec) {
+    Set<String> result = Sets.newHashSet();
+    for (FlowSpecification flowSpec : appSpec.getFlows().values()) {
+      result.addAll(streamsUsedBy(flowSpec));
+    }
+    result.addAll(appSpec.getStreams().keySet());
+    return result;
+  }
+
+  protected final void programListByDataAccess(HttpRequest request, HttpResponder responder,
+                                               Store store, DatasetFramework dsFramework,
+                                               ProgramType type, Data data, String name) {
+    try {
+      if (name.isEmpty()) {
+        responder.sendString(HttpResponseStatus.BAD_REQUEST, data.prettyName().toLowerCase() + " name is empty");
+        return;
+      }
+      String accountId = getAuthenticatedAccountId(request);
+      Id.Program programId = Id.Program.from(accountId, "", "");
+      List<ProgramRecord> programRecords = listProgramsByDataAccess(store, dsFramework, programId, type, data, name);
+      if (programRecords == null) {
+        responder.sendStatus(HttpResponseStatus.NOT_FOUND);
+      } else {
+        responder.sendJson(HttpResponseStatus.OK, programRecords);
+      }
+    } catch (SecurityException e) {
+      responder.sendStatus(HttpResponseStatus.UNAUTHORIZED);
+    } catch (Throwable e) {
+      LOG.error("Got exception:", e);
+      responder.sendStatus(HttpResponseStatus.INTERNAL_SERVER_ERROR);
+    }
+  }
+
+  /**
+   * @return list of program records, an empty list if no programs were found, or null if the stream or
+   * dataset does not exist
+   */
+  private List<ProgramRecord> listProgramsByDataAccess(Store store, DatasetFramework dsFramework,
+                                                       Id.Program programId, ProgramType type,
+                                                       Data data, String name) throws Exception {
+    // search all apps for programs that use this
+    List<ProgramRecord> result = Lists.newArrayList();
+    Collection<ApplicationSpecification> appSpecs = store.getAllApplications(
+      new Id.Namespace(programId.getNamespaceId()));
+    if (appSpecs != null) {
+      for (ApplicationSpecification appSpec : appSpecs) {
+        if (type == ProgramType.FLOW) {
+          for (FlowSpecification flowSpec : appSpec.getFlows().values()) {
+            if ((data == Data.DATASET && usesDataSet(flowSpec, name))
+              || (data == Data.STREAM && usesStream(flowSpec, name))) {
+              result.add(makeProgramRecord(appSpec.getName(), flowSpec, ProgramType.FLOW));
+            }
+          }
+        } else if (type == ProgramType.PROCEDURE) {
+          for (ProcedureSpecification procedureSpec : appSpec.getProcedures().values()) {
+            if (data == Data.DATASET && procedureSpec.getDataSets().contains(name)) {
+              result.add(makeProgramRecord(appSpec.getName(), procedureSpec, ProgramType.PROCEDURE));
+            }
+          }
+        } else if (type == ProgramType.MAPREDUCE) {
+          for (MapReduceSpecification mrSpec : appSpec.getMapReduce().values()) {
+            if (data == Data.DATASET && mrSpec.getDataSets().contains(name)) {
+              result.add(makeProgramRecord(appSpec.getName(), mrSpec, ProgramType.MAPREDUCE));
+            }
+          }
+        }
+      }
+    }
+    if (!result.isEmpty()) {
+      return result;
+    }
+    // if no programs were found, check whether the data exists, return [] if yes, null if not
+    boolean exists = false;
+    if (data == Data.DATASET) {
+      exists = dsFramework.hasInstance(name);
+    } else if (data == Data.STREAM) {
+      exists = store.getStream(new Id.Namespace(Constants.DEFAULT_NAMESPACE), name) != null;
+    }
+    return exists ? result : null;
+  }
+
+  private static boolean usesDataSet(FlowSpecification flowSpec, String dataset) {
+    for (FlowletDefinition flowlet : flowSpec.getFlowlets().values()) {
+      if (flowlet.getDatasets().contains(dataset)) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  private static boolean usesStream(FlowSpecification flowSpec, String stream) {
+    for (FlowletConnection con : flowSpec.getConnections()) {
+      if (FlowletConnection.Type.STREAM == con.getSourceType() && stream.equals(con.getSourceName())) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  /* -----------------  helpers to return Json consistently -------------- */
+
+  protected final static DatasetRecord makeDataSetRecord(String name, String classname) {
+    return new DatasetRecord("Dataset", name, name, classname);
+  }
+
+  protected final static StreamRecord makeStreamRecord(String name, StreamSpecification specification) {
+    return new StreamRecord("Stream", name, name, GSON.toJson(specification));
   }
 }
